question_4_genai_cda/requirements.txt
pandas
python-dotenv
langchain
langchain-core
langchain-google-genai
langchain-experimental
pydantic